Improve requestPost() parameters validation
Parameter validation for posts/comment  are way too restrictive. Let's make is less error prone.

`author` and `body` are mandatory
For a root post (i.e. `title` is not empty):
- `permlink` can be empty
- `parent_author and `parent_permlink` should (must) be empty
- `json_metadata` should at least contains one tag

For a comment (i.e. `title` is empty):
- `permlink` should (must) be empty
- `parent_author and `parent_permlink` cannot be empty
- `json_metadata` can be empty

It would be nice to rename parameters to match name found in the blockchain